refactor: rewrite galore_projector on tensorly update_gap_scheduler goes to new file update_gap_scheduler.py
